feat: load satellite context - remove duplicated parameters code
