+++ src/services/clear.services.ts
@@ -1,16 +1,11 @@
 import {deleteAssets, listCustomDomains, setCustomDomains} from '@junobuild/admin';
 import {deleteAsset} from '@junobuild/core';
 import ora from 'ora';
-import {readJunoConfig} from '../configs/juno.config';
 import {DAPP_COLLECTION} from '../constants/constants';
-import {configEnv} from '../utils/config.utils';
-import {satelliteParameters} from '../utils/satellite.utils';
+import {assertConfigAndLoadSatelliteContext} from '../utils/satellite.utils';
 
 export const clear = async (args?: string[]) => {
-  const env = configEnv(args);
-  const {satellite: satelliteConfig} = await readJunoConfig(env);
-
-  const satellite = await satelliteParameters({satellite: satelliteConfig, env});
+  const {satellite} = await assertConfigAndLoadSatelliteContext(args);
 
   const spinner = ora('Clearing app assets...').start();
 
@@ -38,10 +33,7 @@ const cleanFullPath = (fullPath: string): string => {
 };
 
 export const clearAsset = async ({fullPath, args}: {fullPath: string; args?: string[]}) => {
-  const env = configEnv(args);
-  const {satellite: satelliteConfig} = await readJunoConfig(env);
-
-  const satellite = await satelliteParameters({satellite: satelliteConfig, env});
+  const {satellite} = await assertConfigAndLoadSatelliteContext(args);
 
   const spinner = ora(`Clearing ${fullPath}...`).start();
