+++ src/services/deploy/deploy.assert.services.ts
@@ -1,23 +1,18 @@
 import {satelliteMemorySize, satelliteVersion} from '@junobuild/admin';
 import {yellow} from 'kleur';
 import {compare} from 'semver';
-import {readJunoConfig} from '../../configs/juno.config';
 import {MEMORY_HEAP_WARNING, MEMORY_SIZE_ENDPOINT_VERSION} from '../../constants/deploy.constants';
-import {configEnv} from '../../utils/config.utils';
 import {NEW_CMD_LINE, confirmAndExit} from '../../utils/prompt.utils';
-import {satelliteParameters} from '../../utils/satellite.utils';
+import {assertConfigAndLoadSatelliteContext} from '../../utils/satellite.utils';
 
 export const assertSatelliteMemorySize = async (args?: string[]) => {
-  const env = configEnv(args);
-  const {satellite: satelliteConfig} = await readJunoConfig(env);
+  const {satellite, satelliteConfig} = await assertConfigAndLoadSatelliteContext(args);
 
   const {assertions} = satelliteConfig;
 
   if (assertions?.heapMemory === false) {
     return;
   }
-
-  const satellite = await satelliteParameters({satellite: satelliteConfig, env});
 
   const currentVersion = await satelliteVersion({
     satellite
